+++ tests/test_task.py
@@ -1574,6 +1574,20 @@ class TaskGraphTests(unittest.TestCase):
                 False),
             expected_result_dict)
 
+    def test_duplicate_task_hang_on_exit(self):
+        """TaskGraph: ensure duplicate tasks don't cause taskgraph to hang."""
+        task_graph = taskgraph.TaskGraph(self.workspace_dir, 1)
+        target_path = os.path.join(self.workspace_dir, 'target.txt')
+        content = 'test'
+        for _ in range(10):
+            task = task_graph.add_task(
+                func=_create_file,
+                args=(target_path, content),
+                target_path_list=[target_path],
+                task_name='create content')
+        task_graph.join()
+        task_graph.close()
+
 
 def Fail(n_tries, result_path):
     """Create a function that fails after ``n_tries``."""